array pop_out() now takes t_CKINT instead of t_CKUINT

+++ src/core/util_sndfile.c
@@ -15470,7 +15470,8 @@ static void Fast_Autocorrelation (
 		for (i = k; i < 160; ++i) L_temp2 += sf[i] * sfl[i];
 		f_L_ACF[k] = L_temp2;
 	}
-	scale = MAX_LONGWORD / f_L_ACF[0];
+    // 1.4.2.0 (ge) | added explicit cast to MAX_LONGWORD to remove compiler warning
+	scale = (float)MAX_LONGWORD / f_L_ACF[0];
 
 	for (k = 0; k <= 8; k++) {
 		L_ACF[k] = f_L_ACF[k] * scale;
